+++ vue2-vite4/src/views/manage/File.vue
@@ -4,7 +4,6 @@
     <el-button style="margin-left: 5px" type="primary" @click="load">搜索</el-button>
     <el-button style="margin-left: 5px" type="warning" @click="reset">重置</el-button>
     <div style="margin: 10px 0">
-
       <el-button type="primary" @click="handUpload" style="display: inline-block; margin-right: 5px">上传文件 <i class="el-icon-top"></i></el-button>
       <el-popconfirm
           confirm-button-text='确定'
@@ -16,7 +15,6 @@
       >
         <el-button type="danger" slot="reference" style="margin-right: 5px">批量删除 <i class="el-icon-remove-outline"></i></el-button>
       </el-popconfirm>
-
     </div>
     <el-table :data="tableData" :header-cell-class-name="headerBg" @selection-change="handleSelectionChange" border stripe>
       <el-table-column
@@ -149,6 +147,8 @@ export default {
         if (res.code === '200') {
           this.tableData = res.data.records
           this.total = res.data.total
+        }else {
+          this.$message.error(res.msg)
         }
 
       })
@@ -195,6 +195,10 @@ export default {
     },
     handleDeleteBatch() {
       let ids = this.multipleSelection.map(v => v.id)
+      if(ids.length === 0){
+        this.$message.error("请选择要删除的数据")
+        return
+      }
       fileApi.deleteBatch(ids).then(res => {
         if (res.code === '200') {
           this.$message.success("批量删除成功");
